Created a new utils directory, moved some files, and made updates where needed.
