+++ TestRail/TestRailClient.cs
@@ -1,11 +1,12 @@
-﻿using System;
+﻿using Newtonsoft.Json.Linq;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net;
 using System.Text;
-using Newtonsoft.Json.Linq;
 using TestRail.Types;
+using TestRail.Utils;
 
 namespace TestRail
 {
