+++ TestRail/Types/ConfigurationGroup.cs
@@ -1,5 +1,6 @@
-﻿using System.Collections.Generic;
-using Newtonsoft.Json.Linq;
+﻿using Newtonsoft.Json.Linq;
+using System.Collections.Generic;
+using TestRail.Utils;
 
 namespace TestRail.Types
 {
